requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
coverage==7.6.4
distlib==0.3.9
factory_boy==3.3.1
Faker==30.8.2
fastapi==0.115.4
filelock==3.16.1
flake8==7.1.1
Flake8-pyproject==1.2.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
loguru==0.7.2
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
pycodestyle==2.12.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyflakes==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
virtualenv==20.27.1
